add better grpc test

+++ run-tracetesting-tests.sh
@@ -1,12 +1,16 @@
 #!/bin/sh
 
 STOP=yes
+RESTART=yes
 
 while [[ $# -gt 0 ]]; do
   case $1 in
     --no-stop)
       STOP=no
       shift
+      ;;
+    --no-restart)
+      RESTART=no
       shift
       ;;
     -*|--*)
@@ -18,7 +22,9 @@ while [[ $# -gt 0 ]]; do
 done
 
 
-docker compose -f docker-compose.yaml down
+if [ "$RESTART" == "yes" ]; then
+  docker compose -f docker-compose.yaml down
+fi
 docker compose -f docker-compose.yaml up -d --build --remove-orphans
 docker compose -f docker-compose.yaml -f local-config/docker-compose.testrunner.yaml build
 docker compose -f docker-compose.yaml -f local-config/docker-compose.testrunner.yaml run testrunner
